Implement media-progress() function
As introduced in https://drafts.csswg.org/css-values-5/#media-progress-func
The media-progress() functional notation returns a <number> value
representing current value of the specified media query as a progress value
between two explicit values of the media query.

Note: now it's an editor's draft with a very strong chances to be accepted.
Change-Id: I6bc2ae310d96ba45693e77aa76b15627b4cd66ce