wire up primary font load timings
This CL wires up the load timings of a web font into the text LCP
record as the LCP breakdown timings of that text element.
When a text node is painted, it is enqueued as an LCP candidate.
At this place We check if the applied fonts of the text node contains
the primary font of the fallback font list.
If it does, it means the primary font is applied. Then if the primary
font is a web font, we retrieve the load timings of the web font and
pass them to the text record.
The load timings of the web font is set to a newly added member field
of the CustomFontData object when the resource fetching finish is
notified to the ResourceClient RemoteFontFaceSource object.

Bug: 40276574
Change-Id: Ic13b9d882a97ce0b5101a3cf4693ca1ce2b5ceac