[PEPC] Correctly clamp negative margins to 0
Update the PEPC style for margins to always use the NonNegative length
range.

Also updated the Length equality operator to take into account the
range when deciding equality of computed values as this code reuses the
same expression ref and only updates the length range which was not
being detected as being a different length.

Bug: 1462930
Change-Id: I0e65c4ab3cd40474a675daa964cdae3029b288cc