CloseWatcher: fix Esc key tests
They were interfering with each other by being in the same file. This
meant that most of the tests weren't really testing anything: which is
good, because the keyup and keypress tests had the opposite of the
correct expectations.

This splits up all the tests, and fixes keyup.html and keypress.html
to expect that such listeners are NOT able to prevent the close request.

Bug: 40054591
Change-Id: Id294cb9f1c6e91bec284e5a09042026e54538664