[css-scroll-snap-2] Fix targeted snap point selection
For a page's main frame, Document::SetCSSTarget is called too late to
let the snap containers know which areas to consider targeted. (So, we
need to queue these snap containers up for an update).
For iframes, it is early enough, which is why the tests that were added
to verify snap point selection with targeted elements pass.

This patch adds a new test, prefer-targeted-element-main-frame.html,
which tests snap point selection with a targeted element without iframes
by opening a new window with a hash fragment and reading the result of
the test in that window via `fetch`.

Bug: 324916797
Change-Id: If2f1d0e27368587d5207dc11540fe705d4ab2d36