DOM: Implement Observable#from() conversion operator
This CL implements 3/4 of the Observable#from() conversion operator.
This operator [1], [2] (not quite yet specified) takes a Web IDL `any`
value and converts it to an Observable with the following semantics,
based on what the `any` value actually is under the hood:

  1. Observable: just return the same Observable.
  2. Async iterable: create & return an Observable that, on
     subscription, pulls values from the async iterable, and pushes them
     to the subscriber.
  3. Iterable: create & return an Observable that, on subscription,
     synchronously pushes each value to the subscriber.
  4. Promise: create & return an Observable that, on subscription,
     reacts [3] to given promise with two handlers:
       a. Fulfillment handler: pushes the value to the subscriber, and
          calls complete()
       b. Rejection handler: Calls the subscriber's error with the given
          rejection value

This CL implements (1), (3), and (4) by introducing new subscriber
delegates that are specific to these cases. Async iterable support, (2),
will come in a future CL, as there is not yet equivalent bindings
support for this, so it will need a more thoughtful design.

[1]:
https://github.com/wicg/observable?tab=readme-ov-file#constructing--converting-objects-to-observables
[2]: https://wicg.github.io/observable/#observable-from
[3]: https://webidl.spec.whatwg.org/#dfn-perform-steps-once-promise-is-settled

Bug: 40282760
Change-Id: I6b392ba77221fb1a77e334aa5f151ed6518419ed