Add support for auto keyword in CSS calc-size().
This adds support for the auto keyword in calc-size() by adding the
small changes needed through the whole pipeline: parsing, storage, and
layout.  The layout change here is only a single change to activate the
changes in the dependent CLs.

This adds a number of tests for auto.  The existing tests for animation
from a neutral keyframe to calc-size(auto, size) do not yet work, but
animation between calc-size(auto, ...) values works (as shown by the new
tests).

Bug: 313072
Change-Id: Ibe0c1597eef0bf442f58375da70ea5a12b620acd