Block or allow all MIDI using the existing SysEx permission
We have changed our approach for gating all MIDI access behind a
permission prompt.  Now, the existing SysEx permission and prompt will
be used to control all access to the Web MIDI API.

This CL does the following:
- Roll back registration of the basic MIDI content setting and
  request type
- Move the feature flag location from content/public/common to
  thrid_party/blink/public/common, since the feature is no longer
  related to content
- Update test_driver tests (including WPTs) to request the
  permission with the SysEx flag set to true
- Always show the SysEx version of the prompt by modifying
  midi_access_initializer.cc

Note that one external WPT is also being updated.  Although other
browsers may not need the sysex flag set to true for the idlharness
tests to pass, setting it to true should not cause the tests to fail
since it is stronger than the basic midi permission as per spec:
https://webaudio.github.io/web-midi-api/#permissions-integration

Bug:1420307
Change-Id: I5a6c45641c440f34bfdba0fb2076ae030528c634
Include-Ci-Only-Tests: true